Rename stageExtraTestHeleprs to lessonExtrasTestHelpers
