+++ apps/test/unit/code-studio/components/lessonExtrasTestHelpers.js
